+++ BE/mmmm/src/main/java/com/beautifulyomin/mmmm/domain/stock/dto/response/StockResponseDto.java
@@ -18,5 +18,5 @@ public class StockResponseDto implements StockResponse {
     private BigDecimal priceChange; //전일 대비 금액
     private BigDecimal priceChangeRate; //전일 대비율
     private BigInteger marketCapitalization; //시가총액
-    private BigInteger tradingVolume; //누적 거래량
+    private Long tradingVolume; //누적 거래량
 }
